+++ include/AstSdl.h
@@ -1,6 +1,6 @@
 /*
  * ASTU - AST Utilities
- * A collection of Utilities for Advanced Software Techniques (AST).
+ * A collection of Utilities for Applied Software Techniques (AST).
  * 
  * Copyright (c) 2020 Roman Divotkey, Nora Loimayr. All rights reserved.
  *